% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiomics-package.R
\docType{package}
\name{epiomics-package}
\alias{epiomics}
\alias{epiomics-package}
\title{epiomics: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
