% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data with multiple exposures, multiple outcomes,}
\format{
An dataframe with multiple exposures, outcomes, and omics features.
}
\usage{
data(example_data)
}
\description{
Example data with multiple exposures, multiple outcomes,
}
\examples{
data(example_data)

}
\keyword{datasets}
