% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meet_in_middle.R
\name{meet_in_middle}
\alias{meet_in_middle}
\title{Perform 'omics wide association study}
\usage{
meet_in_middle(
  df,
  exposure,
  outcome,
  omics,
  covars = NULL,
  outcome_model_type = "linear",
  confidence_level = 0.95
)
}
\arguments{
\item{df}{Dataframe}

\item{exposure}{Name of the exposure of interest. Can be either continuous or
dichotomous.}

\item{outcome}{Name of the outcome of interest. Can be either continuous or
dichotomous. For dichotomous variables, must set \code{outcome_model} to
"logistic", and values must be either 0/1 or a factor with the first level
representing the reference group.}

\item{omics}{Names of all omics features in the dataset}

\item{covars}{Names of covariates (can be NULL)}

\item{outcome_model_type}{"linear" for linear models (via lm) or "logistic" for
logistic (via glm)}

\item{confidence_level}{Confidence level for marginal significance
(defaults to 0.95)}
}
\value{
A list of two, for the exposure-omics and omics-outcome MWAS. Each element
contains a data frame with 6 columns:
feature_name: name of the omics feature
estimate: the model estimate for the feature. For linear models, this is the
beta; for logistic models, this is the log odds.
se: Standard error of the estimate
p_value: p-value for the estimate
adjusted_pval: FDR adjusted p-value
threshold: Marginal significance, based on unadjusted p-values
}
\description{
Implements a meet in the middle analysis for identifying omics associated
with both exposures and outcomes
}
