% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owas.R
\name{owas}
\alias{owas}
\title{Perform 'omics wide association study}
\usage{
owas(
  df,
  var,
  omics,
  covars = NULL,
  var_exposure_or_outcome,
  family = "gaussian",
  confidence_level = 0.95,
  conf_int = FALSE
)
}
\arguments{
\item{df}{Dataset}

\item{var}{Name of the variable or variables of interest- this is usually
either an exposure variable or an outcome variable. Can be either
continuous or dichotomous. For dichotomous variables, must set \code{family}
to "binomial", and values must be either 0/1 or a factor with the first
level representing the reference group. Can handle multiple variables, but
they must all be of the same \code{family}.}

\item{omics}{Names of all omics features in the dataset}

\item{covars}{Names of covariates (can be NULL)}

\item{var_exposure_or_outcome}{Is the variable of interest an exposure
(independent variable) or outcome (dependent variable)? Must be either
"exposure" or "outcome"}

\item{family}{"gaussian" (defualt) for linear models (via lm) or "binomial" for
logistic (via glm)}

\item{confidence_level}{Confidence level for marginal significance
(defaults to 0.95, or an alpha of 0.05)}

\item{conf_int}{Should Confidence intervals be generated for the estimates?
Default is FALSE. Setting to TRUE will take longer. For logistic models,
calculates Wald confidence intervals via \code{confint.default}.}
}
\value{
A data frame with 6 columns:
feature_name: name of the omics feature
estimate: the model estimate for the feature. For linear models, this is the
beta; for logistic models, this is the log odds.
se: Standard error of the estimate
test statistic: t-value
p_value: p-value for the estimate
adjusted_pval: FDR adjusted p-value
threshold: Marginal significance, based on unadjusted p-values
}
\description{
Implements an omics wide association study with the option of using the
'omics data as either the dependent variable (i.e., for performing an
exposure --> 'omics analysis) or using the 'omics as the independent variable
(i.e., for performing an 'omics --> outcome analysis). Allows for either
continuous or dichotomous outcomes, and provides the option to adjust for
covariates.
}
\examples{
# Simulate dataset
set.seed(4656)
n_omic_ftrs = 100
n_ids = 400
# Simulate omics
omics_df <- matrix(nrow = n_ids, 
                   ncol = n_omic_ftrs)
omics_df <- apply(omics_df, MARGIN = 2, FUN = function(x){rnorm(n_ids)})
omics_df <- as.data.frame(omics_df)
colnames(omics_df) <- paste0("feature_", colnames(omics_df))
# Simulate covariates and outcomes
cov_out <- data.frame(id = c(1:n_ids),
                      sex = sample(c("male", "female"),
                                   n_ids, replace=TRUE,prob=c(.5,.5)),
                      age = rnorm(10, 10, 2),
                      weight =  rlnorm(n_ids, meanlog = 3, sdlog = 0.2),
                      exposure1 = rlnorm(n_ids, meanlog = 2.3, sdlog = 1),
                      exposure2 = rlnorm(n_ids, meanlog = 2.3, sdlog = 1),
                      exposure3 = rlnorm(n_ids, meanlog = 2.3, sdlog = 1),
                      disease1 = sample(0:1, n_ids, replace=TRUE, prob=c(.9,.1)), 
                      disease2 = sample(0:1, n_ids, replace=TRUE,prob=c(.9,.1)))

# Create Test Data
test_data <- cbind(cov_out, omics_df)

# Get names of omics
colnames_omic_fts <- colnames(test_data)[grep("feature_",
                                              colnames(test_data))]

# Get names of exposures
expnms = c("exposure1", "exposure2", "exposure3")

# Run function with one continuous exposure as the variable of interest
owas(df = test_data, 
     var = "exposure1", 
     omics = colnames_omic_fts, 
     covars = c("age", "sex"), 
     var_exposure_or_outcome = "exposure", 
     family = "gaussian")
     
# Run function with multiple continuous exposures as the variable of interest
owas(df = test_data, 
     var = expnms, 
     omics = colnames_omic_fts, 
     covars = c("age", "sex"), 
     var_exposure_or_outcome = "exposure", 
     family = "gaussian")

# Run function with dichotomous outcome as the variable of interest
owas(df = test_data, 
     var = "disease1", 
     omics = colnames_omic_fts, 
     covars = c("age", "sex"), 
     var_exposure_or_outcome = "outcome", 
     family = "binomial")
 

}
