% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owas_qgcomp.R
\name{owas_qgcomp}
\alias{owas_qgcomp}
\title{Perform 'omics wide association study using qgcomp}
\usage{
owas_qgcomp(df, expnms, omics, covars = NULL, q = 4, confidence_level = 0.95)
}
\arguments{
\item{df}{Dataset}

\item{expnms}{Name of the exposures. Can be either continuous or
dichotomous. For dichotomous variables, must set \code{q} to
"NULL", and values must be either 0/1.}

\item{omics}{Names of all omics features in the dataset}

\item{covars}{Names of covariates (can be NULL)}

\item{q}{NULL or number of quantiles used to create quantile indicator variables representing the exposure variables. Defaults to 4If NULL, then gcomp proceeds with un-transformed version of exposures in the input datasets (useful if data are already transformed, or for performing standard g-computation).}

\item{confidence_level}{Confidence level for marginal significance
(defaults to 0.95, or an alpha of 0.05)}
}
\value{
A data frame with 6 columns:
feature_name: name of the omics feature
psi: the model estimate for the feature. For linear models, this is the
beta; for logistic models, this is the log odds.
se: Standard error of the estimate
p_value: p-value for the estimate
adjusted_pval: FDR adjusted p-value
threshold: Marginal significance, based on unadjusted p-values
}
\description{
Implements an omics wide association study using QGComp to model associations of
exposure mixtures with each individual 'omics feature as an outcome
'omics data as either the dependent variable. Allows for either
continuous or dichotomous outcomes, and provides the option to adjust for
covariates.
}
\examples{
# Simulate dataset
set.seed(4656)
n_omic_ftrs = 100
n_ids = 400
# Simulate omics
omics_df <- matrix(nrow = n_ids, 
                   ncol = n_omic_ftrs)
omics_df <- apply(omics_df, MARGIN = 2, FUN = function(x){rnorm(n_ids)})
omics_df <- as.data.frame(omics_df)
colnames(omics_df) <- paste0("feature_", colnames(omics_df))
# Simulate covariates and outcomes
cov_out <- data.frame(id = c(1:n_ids), 
                      sex = sample(c("male", "female"), 
                                  n_ids, replace=TRUE,prob=c(.5,.5)),
                     age = rnorm(10, 10, 2),
                     exposure1 = rlnorm(n_ids, meanlog = 2.3, sdlog = 1),
                     exposure2 = rlnorm(n_ids, meanlog = 2.3, sdlog = 1),
                     exposure3 = rlnorm(n_ids, meanlog = 2.3, sdlog = 1),
                     disease = sample(0:1, n_ids, replace=TRUE,prob=c(.9,.1)),
                     weight =  rlnorm(n_ids, meanlog = 3, sdlog = 0.2))

# Create Test Data
test_data <- cbind(cov_out, omics_df)

# Get names of omics
colnames_omic_fts <- colnames(test_data)[grep("feature_",
                                              colnames(test_data))]

# Names of exposures in mixture
 exposure_names = c("exposure1", "exposure2", "exposure3")

# Run function without covariates
out <- owas_qgcomp(df = test_data,
                   expnms = exposure_names,
                   omics = colnames_omic_fts,
                   q = 4, 
                   confidence_level = 0.95) 


# Run analysis with covariates
out <- owas_qgcomp(df = test_data,
                   expnms = c("exposure1", "exposure2", "exposure3"),
                   covars = c("weight", "age", "sex"),
                   omics = colnames_omic_fts,
                   q = 4, 
                   confidence_level = 0.95) 
 

}
