% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiomics-package.R
\docType{package}
\name{epiomics-package}
\alias{epiomics}
\alias{epiomics-package}
\title{epiomics: Functions for Analyzing 'Omics Data in Observational Studies}
\description{
Provides a collection of fast and flexible functions for the analysis of 'omics data in observational studies.
}
\author{
\strong{Maintainer}: Jesse Goodrich \email{jagoodri@usc.edu} (\href{https://orcid.org/0000-0001-6615-0472}{ORCID})

}
\keyword{internal}
