% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meet_in_middle.R
\name{meet_in_middle}
\alias{meet_in_middle}
\title{Perform 'omics wide association study}
\usage{
meet_in_middle(
  df,
  exposure,
  outcome,
  omics,
  covars = NULL,
  outcome_family = "gaussian",
  confidence_level = 0.95,
  conf_int = FALSE
)
}
\arguments{
\item{df}{Dataframe}

\item{exposure}{Name of the exposure of interest. Can be either continuous or
dichotomous.}

\item{outcome}{Name of the outcome of interest. Can be either continuous or
dichotomous. For dichotomous variables, must set \code{outcome_family} to
"logistic", and values must be either 0/1 or a factor with the first level
representing the reference group.}

\item{omics}{Names of all omics features in the dataset}

\item{covars}{Names of covariates (can be NULL)}

\item{outcome_family}{"gaussian" for linear models (via lm) or "binomial" for
logistic (via glm)}

\item{confidence_level}{Confidence level for marginal significance
(defaults to 0.95)}

\item{conf_int}{Should Confidence intervals be generated for the estimates?
Default is FALSE. Setting to TRUE will take longer. For logistic models,
calculates Wald confidence intervals via \code{confint.default}.}
}
\value{
A list of three dataframes, containing:
\enumerate{
\item Results from the Exposure-Omics Wide Association Study
\item Results from the Omics-Outcome Wide Association Study
\item Overlapping significant features from 1 and 2.
For each omics wide association, results are provided in a data frame with 6 columns:
feature_name: name of the omics feature
estimate: the model estimate for the feature. For linear models, this is the
beta: for logistic models, this is the log odds.
se: Standard error of the estimate
p_value: p-value for the estimate
adjusted_pval: FDR adjusted p-value
threshold: Marginal significance, based on unadjusted p-values
}
}
\description{
Implements a meet in the middle analysis for identifying omics associated
with both exposures and outcomes, as described by
}
\examples{
 # Simulate dataset
set.seed(4656)
n_omic_ftrs = 100
n_ids = 400
# Simulate omics
omics_df <- matrix(nrow = n_ids,
                   ncol = n_omic_ftrs)
omics_df <- apply(omics_df, MARGIN = 2, FUN = function(x){rnorm(n_omic_ftrs)})
omics_df <- as.data.frame(omics_df)
colnames(omics_df) <- paste0("feature_", colnames(omics_df))
# Simulate covariates and outcomes
cov_out <- data.frame(id = c(1:n_ids),
                      sex = sample(c("male", "female"),
                                   n_ids, replace=TRUE,prob=c(.5,.5)),
                      age = rnorm(10, 10, 2),
                      exposure = rlnorm(n_ids, meanlog = 2.3, sdlog = 1),
                      disease = sample(0:1, n_ids, replace=TRUE,prob=c(.9,.1)),
                      weight =  rlnorm(n_ids, meanlog = 3, sdlog = 0.2))

# Create Test Data
test_data <- cbind(cov_out, omics_df)

# Get names of omics
colnames_omic_fts <- colnames(test_data)[grep("feature_",
                                              colnames(test_data))]

# Meet in the middle with a dichotomous outcome
res <- meet_in_middle(df = test_data,
                      exposure = "exposure", 
                      outcome = "disease", 
                      omics = colnames_omic_fts,
                      covars = c("age", "sex"), 
                      outcome_family = "binomial")

# Meet in the middle with a continuous outcome 
res <- meet_in_middle(df = test_data,
                      exposure = "exposure", 
                      outcome = "weight", 
                      omics = colnames_omic_fts,
                      covars = c("age", "sex"), 
                      outcome_family = "gaussian")

# Meet in the middle with a continuous outcome and no covariates
res <- meet_in_middle(df = test_data,
                      exposure = "exposure", 
                      outcome = "weight", 
                      omics = colnames_omic_fts,
                      outcome_family = "gaussian")

}
